# old/api/docs/swagger.yaml
definitions:
...
  github_com_nicolailuther_butter_internal_models.Content:
    properties:
      accounts:
        items:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Account'
        type: array
      avg_views:
        type: integer
      client_id:
        type: integer
      content_files:
        items:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentFile'
        type: array
      created_at:
        type: string
      enabled:
        type: boolean
      folder:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentFolder'
      folder_id:
        description: Relation to organize content
        type: integer
      id:
        type: integer
      last_generated_at:
        type: string
      name:
        type: string
      times_generated:
        type: integer
      times_posted:
        description: Generation and posting stats
        type: integer
      total_views:
        type: integer
      type:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.ContentType'
      updated_at:
        type: string
      use_mirror:
        description: Default generation settings
        type: boolean
      use_overlays:
        type: boolean
    type: object
...
  github_com_nicolailuther_butter_internal_models.ContentFolder:
    properties:
      children:
        items:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentFolder'
        type: array
      client_id:
        type: integer
      contents:
        items:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Content'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        description: Remove default:null, use pointer instead
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_responses.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: Default
    url: https://default.dev
  description: Backend service for Butter
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Butter API
  version: "1.0"
paths:
...
  /api/v2/content:
    get:
      description: Retrieve folder contents by client and folder ID
      parameters:
      - description: Client ID to filter contents
        in: query
        name: client_id
        required: true
        type: integer
      - description: Folder ID to get contents from (0 for root)
        in: query
        name: folder_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Folder with contents retrieved successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentFolder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get content folder and contents
      tags:
      - content-v2
    post:
      consumes:
      - multipart/form-data
      description: Create new content with file uploads
      parameters:
      - description: Client ID
        in: formData
        name: client_id
        required: true
        type: integer
      - description: Content type
        in: formData
        name: type
        required: true
        type: string
      - description: Folder ID (0 for root)
        in: formData
        name: folder_id
        type: integer
      - description: Content files to upload
        in: formData
        name: content_files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Content created successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Content'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create content
      tags:
      - content-v2
  /api/v2/content/{id}:
    delete:
      description: Delete existing content
      parameters:
      - description: Content ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Content deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete content
      tags:
      - content-v2
    patch:
      consumes:
      - application/json
      description: Update existing content
      parameters:
      - description: Content ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Content update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.UpdateContent'
      produces:
      - application/json
      responses:
        "200":
          description: Content updated successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Content'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update content
      tags:
      - content-v2
  /api/v2/content/{id}/assignations/{account}:
    delete:
      description: Unassign an account from content
      parameters:
      - description: Content ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account ID
        in: path
        name: account
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Content updated successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Content'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unassign Account from Content
      tags:
      - content-v2
    patch:
      consumes:
      - application/json
      description: Update an existing content account assignment
      parameters:
      - description: Content ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account ID
        in: path
        name: account
        required: true
        type: integer
      - description: Content account update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.UpdateContentAccount'
      produces:
      - application/json
      responses:
        "200":
          description: Content account assignment updated successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Content account assignment not found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Content Account Assignment
      tags:
      - content-v2
    post:
      description: Assign an account to content
      parameters:
      - description: Content ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account ID
        in: path
        name: account
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Content updated successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Content'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign Account to Content
      tags:
      - content-v2
  /api/v2/content/by-account/{account}:
    get:
      description: Get all contents assigned to a specific account
      parameters:
      - description: Account ID
        in: path
        name: account
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of contents assigned to account
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentAccount'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Contents by Account
      tags:
      - content-v2
  /api/v2/content/folders:
    post:
      consumes:
      - application/json
      description: Create a new content folder
      parameters:
      - description: Folder creation payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.CreateFolder'
      produces:
      - application/json
      responses:
        "201":
          description: Folder created successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentFolder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create content folder
      tags:
      - content-v2
  /api/v2/content/folders/{id}:
    delete:
      description: Delete an existing content folder (only if empty)
      parameters:
      - description: Folder ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Folder deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Folder not found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "409":
          description: Folder contains content
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete content folder
      tags:
      - content-v2
    patch:
      consumes:
      - application/json
      description: Update an existing content folder
      parameters:
      - description: Folder ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Folder update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.UpdateFolder'
      produces:
      - application/json
      responses:
        "200":
          description: Folder updated successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.ContentFolder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "404":
          description: Folder not found
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update content folder
      tags:
      - content-v2
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PushApiKey:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"